refactor: 데이터 저장 없이 촬영 사진 바로 임시 파일 저장하도록 수정

+++ mirroringBooth/StoreTests/CameraPreviewStoreTests.swift
@@ -14,7 +14,7 @@ final class MockCameraManager: CameraManageable {
     var rawData: ((CMSampleBuffer) -> Void)?
     var onEncodedData: ((Data) -> Void)?
     var onTransferCompleted: (() -> Void)?
-    var onAllPhotosStored: ((Int) -> Void)?
+    var onAllPhotosStored: (() -> Void)?
 
     private(set) var startSessionCallCount = 0
     private(set) var stopSessionCallCount = 0
